Sets up Vitest and React Testing Library
Adds Vitest, React Testing Library, MSW, and jsdom for component testing.

Includes initial tests for ShiftForm and SummaryCard components and configures MSW for API mocking. Also, adds `test`, `test:run`, and `test:ui` scripts to `package.json` for ease of use.

+++ src/lib/validations.ts
@@ -43,7 +43,7 @@ export const shiftFormSchema = z
     shiftDate: z.date(),
     shiftStartTime: z.string(),
     shiftEndTime: z.string(),
-    tips: z.number().min(0).optional(),
+    tips: z.number().int().min(0).optional(),
   })
   .refine(
     (data) => {
